Modified Generate.jl to allow generation of a Codata module for different unit types
I parameterized autogen_codata_module with a function that takes 4 strings
name,value,units,unc for a constant name with value value, units units and
uncentatity unc and returns a string defining a constant for a specific
type. I added a generating function for the SIUnits.jl package. I also addef
a few other things, I made generate header a parameter for autogen_codata_module
to go along with allowing different types, I also allowed the name of the
generated file to be chosen as well as to determine weather or not to export all
the constants.

I made the file into a module and exported the autogen_codata_module function
as well as functions to generate modules for the default configuration and
for the siunits module. I added an internal function that downloads the
codata allascii.txt file to use, I also added said file to the repository which
is necessary since the module needs to be re-created for different types.